requirements.txt
# 파일명: requirements.txt

# Google 번역 서비스를 사용하기 위한 비공식 파이썬 라이브러리입니다.
# 특정 버전(4.0.0-rc1)이 안정성이 높아 명시적으로 지정합니다.
googletrans

# Google Gemini API를 사용하기 위한 공식 라이브러리입니다.
google-generativeai

# OpenAI (ChatGPT) API를 사용하기 위한 공식 라이브러리입니다.
openai

# Anthropic (Claude) API를 사용하기 위한 공식 라이브러리입니다.
anthropic
